ref(crons): Refactor out landing page onboarding panel
